backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
openai
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.4.4
